+++ RC_DPU_PI_200M_4/src/app/user_func.h
@@ -41,7 +41,7 @@ int HoppingEnd(taliseDevice_t * const pd, const uint64_t freq);
 int CHScanStart(const uint8_t CH, const uint8_t ITER_CNT);
 static int IterSpectrum(void);
 static int AdrvGainCtrl(const uint64_t FREQ);
-int BWScanStart(const uint64_t FREQ, uint64_t BW, uint16_t RBW);
+int BWScanStart(const uint64_t FREQ, uint64_t BW, uint32_t RBW);
 
 
 #endif /* SRC_APP_USER_FUNC_H_ */
